Move funcs to encodings module
